refactor: 카테고리 위치 수정
