+++ src/main/java/com/souf/soufwebsite/global/common/category/dto/CategoryReqDto.java
@@ -1,0 +1,8 @@
+package com.souf.soufwebsite.global.common.category.dto;
+
+public record CategoryReqDto(
+        Long firstCategory,
+        Long secondCategory,
+        Long thirdCategory) {
+
+}
