+++ src/main/java/com/souf/soufwebsite/global/common/category/entity/FirstCategory.java
@@ -1,4 +1,4 @@
-package com.souf.soufwebsite.global.common.category;
+package com.souf.soufwebsite.global.common.category.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
